剂量Long改成BigDecimal

+++ doctor-event-api/src/main/java/io/terminus/doctor/event/model/DoctorVaccinationPigWarn.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.Builder;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -79,7 +80,7 @@ public class DoctorVaccinationPigWarn implements Serializable {
     /**
      * 消耗剂量
      */
-    private Long dose;
+    private BigDecimal dose;
     
     /**
      * 备注
